+++ buildh/cli.py
@@ -12,6 +12,7 @@ from . import lipids
 from . import init_dics
 from . import core
 from . import writers
+from . import utils
 
 # For debugging.
 DEBUG = False
@@ -44,75 +45,6 @@ def isfile(path):
     return path
 
 
-def check_slice_options(system, first_frame=None, last_frame=None):
-    """Verify the slicing options and return a range of frames in MDAnalysis.
-
-    This function check whether the first frame and the last frame are consistent
-    within themselves (`first_frame` cant be superior to `last_frame`) and
-    with the trajectory supplied (if the trajectory starts at 1000ps, `first_frame`
-    cant be equal to 0 for example).
-    Then, the function translate the range from picosecond-time to the number of frame
-    in MDanalysis.
-
-    Parameters
-    ----------
-    system : MDAnalysis universe instance
-        This is the universe *without* hydrogen.
-    first_frame : int
-        the first frame to read (in ps)
-    last_frame : int
-        the last frame to read (in ps)
-
-    Return
-    ------
-    tuple of int
-        The number of first and last frame
-
-    Raises
-    ------
-    IndexError
-        When the slicing options are out of range.
-    """
-    # From the trajectory, get the time of the first and last frame
-    traj_first_frame = int(system.trajectory.time)
-    traj_last_frame = int(system.trajectory.time +
-                          system.trajectory.dt * (system.trajectory.n_frames - 1))
-
-    # If no bound is given, take the full trajectory
-    if not first_frame and not last_frame:
-        return (0, system.trajectory.n_frames)
-
-    # If only one bound is given
-    if not first_frame:
-        first_frame = traj_first_frame
-    if not last_frame:
-        last_frame = traj_last_frame
-
-
-    # Check abnormal range
-    if first_frame < 0 or last_frame < 0:
-        raise IndexError("Incorrect slice options.")
-    if first_frame > last_frame:
-        raise  IndexError("Incorrect slice options")
-
-    # Check if the range fits into the range of the trajectory
-    if first_frame < traj_first_frame or last_frame < traj_first_frame:
-        raise  IndexError("Incorrect slice options")
-    if first_frame > traj_last_frame or last_frame > traj_last_frame:
-        raise  IndexError("Incorrect slice options")
-
-    # Translate the time range into a number range.
-    # Find the index of element in the list of frames (in ps) which has the minimum distance
-    # from the first or last frame (in ps) given.
-    frames = np.arange(traj_first_frame, traj_last_frame + 1, int(system.trajectory.dt))
-    number_first_frame = (np.abs(frames - first_frame)).argmin()
-    number_last_frame  = (np.abs(frames -  last_frame)).argmin()
-    # Include last frame into account for slicing by adding 1
-    number_last_frame = number_last_frame + 1
-
-    return (number_first_frame, number_last_frame)
-
-
 def parse_cli():
     """Handle the user parameters from the command line."""
     # Retrieve list of supported lipids
@@ -131,13 +63,13 @@ def parse_cli():
     parser = argparse.ArgumentParser(description=message,
                                      epilog=epilog)
     # Avoid tpr for topology cause there's no .coord there!
-    parser.add_argument("topfile", type=isfile,
-                        help="Topology file (pdb or gro).")
-    parser.add_argument("-x", "--xtc", type=isfile,
-                        help="Input trajectory file in xtc format.")
+    parser.add_argument("-c", "--coord", type=isfile, required=True,
+                        help="Coordinate file (pdb or gro).")
+    parser.add_argument("-t", "--traj", type=isfile,
+                        help="Input trajectory file. Could be in XTC, TRR or DCD format.")
     parser.add_argument("-l", "--lipid", type=str, required=True,
                         help="Residue name of lipid to calculate the OP on (e.g. POPC).")
-    parser.add_argument("-tl", "--lipid_topology", type=isfile, nargs='+',
+    parser.add_argument("-lt", "--lipid_topology", type=isfile, nargs='+',
                         help="User topology lipid json file(s). Mandatory to build hydrogens.")
     parser.add_argument("-d", "--defop", required=True, type=isfile,
                         help="Order parameter definition file. Can be found on "
@@ -161,17 +93,19 @@ def parse_cli():
                         "the value of each lipid and each frame as a matric")
     options = parser.parse_args()
 
-    # Top file is "options.topfile", xtc file is "options.xtc", pdb output file is
-    # "options.pdbout", xtc output file is "options.xtcout".
-    # Check topology file extension.
-    if not options.topfile.endswith("pdb") and not options.topfile.endswith("gro"):
-        parser.error("Topology must be given in pdb or gro format")
 
-    # Append user lipid topologies to the default ones
+    # Check topology file extension.
+    if not options.coord.endswith("pdb") and not options.coord.endswith("gro"):
+        parser.error("Coordinates file must be given in .pdb or .gro format.")
+
+    # Use only the user lipid topologies
     if options.lipid_topology:
-        lipids_files += [pathlib.Path(f) for f in options.lipid_topology]
+        lipids_files = [pathlib.Path(f) for f in options.lipid_topology]
         # Regenerate lipid topologies dictionary
-        lipids_tops = lipids.read_lipids_topH(lipids_files)
+        try:
+            lipids_tops = lipids.read_lipids_topH(lipids_files)
+        except ValueError as e:
+            parser.error(e)
         # Regenerate str list of supported lipids.
         lipids_supported_str = ", ".join(lipids_tops.keys())
 
@@ -181,63 +115,14 @@ def parse_cli():
     try:
         lipids_info = lipids_tops[options.lipid]
     except KeyError:
-        parser.error(f"Lipid {options.lipid} is not supported."
-                     "List of supported lipids are: {lipids_supported_str}")
+        parser.error(f"Lipid {options.lipid} is not supported. "
+                     f"List of supported lipids are: {lipids_supported_str}")
 
     # Slicing only makes sense with a trajectory
-    if not options.xtc and (options.begin or options.end):
+    if not options.traj and (options.begin or options.end):
         parser.error("Slicing is only possible with a trajectory file.")
 
     return options, lipids_info
-
-
-def check_def_file(universe, res_name, atoms_name):
-    """Check if atoms from the definition file are present in the structure in `universe`.
-
-    This function return false if there is one missing in the structure.
-    Print also an error message.
-
-    Parameters
-    ----------
-    universe : MDAnalysis universe instance
-    res_name : str
-        lipid residue name
-    atoms_name : list of str
-        list of atom names
-
-    Returns
-    -------
-    Bool
-        True if all atoms are found in the structure. False otherwise.
-    """
-    for atom_name in atoms_name:
-        if not check_atom(universe, res_name, atom_name):
-            print(f"Atom {atom_name} of residue {res_name} from definition "
-                  "file is not found in your system.")
-            return False
-
-    return True
-
-
-def check_atom(universe, res_name, atom_name):
-    """Check if `atom_name` from residue `res_name` is present in `universe`.
-
-    Parameters
-    ----------
-    universe : MDAnalysis universe instance
-    res_name : str
-        residue name
-    atom_name : str
-        atom name
-
-    Returns
-    -------
-    Bool
-        True if the atom is present. False otherwise.
-    """
-    if len(universe.select_atoms(f"resname {res_name} and name {atom_name}")) == 0:
-        return False
-    return True
 
 
 def main():
@@ -250,31 +135,32 @@ def main():
 
     # 2) Create universe without H.
     print("Constructing the system...")
-    if args.xtc:
+    if args.traj:
         try:
-            universe_woH = mda.Universe(args.topfile, args.xtc)
-            begin, end = check_slice_options(universe_woH, args.begin, args.end)
+            universe_woH = mda.Universe(args.coord, args.traj)
+            begin, end = utils.check_slice_options(universe_woH, args.begin, args.end)
             traj_file = True
         except IndexError:
-            raise UserWarning("Slicing options are not correct.") from None
+            sys.exit("Slicing options are not correct.")
         except:
-            raise UserWarning("Can't create MDAnalysis universe with files {} "
-                              "and {}".format(args.topfile, args.xtc)) from None
+            sys.exit(f"Can't create MDAnalysis universe with files {args.coord} and {args.traj}.")
     else:
         try:
-            universe_woH = mda.Universe(args.topfile)
+            universe_woH = mda.Universe(args.coord)
             begin = 0
             end = 1
             traj_file = False
         except:
-            raise UserWarning("Can't create MDAnalysis universe with file {}"
-                              .format(args.topfile)) from None
+            sys.exit(f"Can't create MDAnalysis universe with file {args.coord}.")
 
 
     # 2) Initialize dic for storing OP.
     # Init dic of correspondance : {('C1', 'H11'): 'gamma1_1',
     # {('C1', 'H11'): 'gamma1_1', ...}.
-    dic_atname2genericname = init_dics.make_dic_atname2genericname(args.defop)
+    try:
+        dic_atname2genericname = init_dics.make_dic_atname2genericname(args.defop)
+    except ValueError as e:
+        sys.exit(e)
     # Initialize dic_OP (see function init_dic_OP() for the format).
     dic_OP, dic_corresp_numres_index_dic_OP = init_dics.init_dic_OP(universe_woH,
                                                                     dic_atname2genericname,
@@ -288,12 +174,16 @@ def main():
 
     # Check if the lipid topology match the the structure.
     if not lipids.check_topology(universe_woH, dic_lipid):
-        sys.exit(f"The topology chosen does not match the structure provided {args.topfile}")
+        sys.exit(f"The topology chosen does not match the structure provided in {args.coord}")
 
     # Check if atoms name in the def file are present in the structure.
     atoms_name = [heavy_atom for (heavy_atom, _) in dic_atname2genericname.keys()]
-    if not check_def_file(universe_woH, dic_lipid['resname'], atoms_name):
-        sys.exit(f"Atoms defined in {args.defop} are missing in the structure {args.topfile}.")
+    if not utils.check_def_file(universe_woH, dic_lipid['resname'], atoms_name):
+        sys.exit(f"Atoms defined in {args.defop} are missing in the structure {args.coord}.")
+
+    # Check the def file and the topology are coherent.
+    if not utils.check_def_topol_consistency(dic_Cname2Hnames, dic_lipid):
+        sys.exit(f"Atoms defined in {args.defop} are not consistent with topology chosen.")
 
 
     print("System has {} atoms".format(len(universe_woH.coord)))
@@ -303,12 +193,13 @@ def main():
     #  with arg -d) needs to contain all possible C-H pairs !!!
     if args.opdbxtc:
 
-        if core.is_allHs_present(args.defop, dic_lipid, dic_Cname2Hnames):
+        if utils.is_allHs_present(args.defop, dic_lipid, dic_Cname2Hnames):
             core.gen_coordinates_calcOP(args.opdbxtc, universe_woH, dic_OP, dic_lipid,
                                         dic_Cname2Hnames, dic_corresp_numres_index_dic_OP,
                                         begin, end, traj_file)
         else:
-            raise UserWarning("Error on the number of H's to rebuild.")
+            sys.exit(f"Error on the number of H's to rebuild. An output trajectory has been "
+                     f"requestest but {args.defop} doesn't contain all hydrogens to rebuild.")
 
     # 6) If no traj output file requested, use fast indexing to speed up OP
     # calculation. The function fast_build_all_Hs() returns nothing, dic_OP